refactor(ContextMenu): Move `useRightClick` hook
